\name{sgpFlowMatrices}
\alias{sgpFlowMatrices}
\docType{data}
\title{Coefficient matrices for sgpFlow analyses}
\description{
An environment containing collections of SGP coefficient matrices for performing sgpFlow analyses. Matrices might be updated periodically.
The environment currently contains matrices associated with the following states/organizations/years:

\describe{
    \item{\code{DEMO_sgpFlowMatrices}}{Contains single-cohort (2023-2024) and super-cohort (2020-2021 to 2023-2024) sgpFlow matrices for Demonstration DEMO}
}
}

\source{Large Scale Summative Assessment Data}
\keyword{datasets}
