\name{sgpFlowMatrices-package}
\alias{sgpFlowMatrices-package}
\docType{package}
\title{Coefficient matrices for sgpFlow analyses}
\description{
	The package sgpFlowMatrices contains coefficient matrices for multiple states/assessment for sgpFlow analyses.
}

\details{
\tabular{ll}{
Package: \tab sgpFlowMatrices\cr
Type: \tab Package\cr
Version: \tab 0.0-0.2\cr
Date: \tab 2024-12-3\cr
License: \tab GPL-3\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
