% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpFlowMatrices.R
\docType{data}
\name{sgpFlowMatrices}
\alias{sgpFlowMatrices}
\title{Coefficient matrices for sgpFlow analyses}
\format{
An environment containing coefficient matrices for different states/organizations/years.
}
\source{
Large Scale Summative Assessment Data
}
\usage{
sgpFlowMatrices
}
\description{
An environment containing collections of SGP coefficient matrices for performing sgpFlow analyses. 
Matrices might be updated periodically.
}
\section{Current Contents}{

The environment currently contains matrices associated with the following states/organizations/years:
\describe{
    \item{\code{DEMO_sgpFlowMatrices}}{Contains single-cohort (2023-2024) and super-cohort (2020-2021 to 2023-2024) sgpFlow matrices for Demonstration DEMO}
    \item{\code{IN_sgpFlowMatrices}}{Contains single-cohort (2024) and super-cohort (2021 to 2024) sgpFlow matrices for Indiana data}
}
}

\keyword{datasets}
